[GEOT-6709] Pull out BaseUnitFormatter to hide implementation details from users
